+++ lizmap/definitions/filter_by_form.py
@@ -1,6 +1,7 @@
 """Definitions for filter by form."""
 
 from lizmap.definitions.base import BaseDefinitions, InputType
+from lizmap.definitions.definitions import LwcVersions
 from lizmap.qgis_plugin_tools.tools.i18n import tr
 
 __copyright__ = 'Copyright 2020, 3Liz'
@@ -26,6 +27,7 @@ class FilterByFormDefinitions(BaseDefinitions):
                 'The title to give to the input, which will be displayed above the form input. '
                 'For example "Choose a category" for a layer field called "category".')
         }
+        # TODO switch to enum for the list with icons
         self._layer_config['type'] = {
             'type': InputType.List,
             'header': tr('Type'),
@@ -39,6 +41,8 @@ class FilterByFormDefinitions(BaseDefinitions):
             'tooltip': tr(
                 'The field name to apply the filter on.')
         }
+        # In a near future, we should merge min_date into start_field
+        # LWC 3.7 is able to read it already
         self._layer_config['min_date'] = {
             'type': InputType.Field,
             'header': tr('Date minimum'),
@@ -46,6 +50,8 @@ class FilterByFormDefinitions(BaseDefinitions):
             'tooltip': tr(
                 'The field containing the start date of your feature (ex: "start_date" of an event).')
         }
+        # In a near future, we should merge max_date into end_field
+        # LWC 3.7 is able to read it already
         self._layer_config['max_date'] = {
             'type': InputType.Field,
             'header': tr('Date maximum'),
@@ -55,6 +61,20 @@ class FilterByFormDefinitions(BaseDefinitions):
                 'one for the start date and another for the end date, you can differentiate them. '
                 'If not, you need to use the same field name for Min date and Max date.'
             )
+        }
+        self._layer_config['start_field'] = {
+            'type': InputType.Field,
+            'header': tr('Start field'),
+            'default': '',
+            'tooltip': tr('The field containing the minimum/start value'),
+            # 'version': LwcVersions.Lizmap_3_7, # This field is kinda not new, used for legacy CFG from < 3.7
+        }
+        self._layer_config['end_field'] = {
+            'type': InputType.Field,
+            'header': tr('End field'),
+            'default': '',
+            'tooltip': tr('The field containing the maximum/end value of your data.'),
+            'version': LwcVersions.Lizmap_3_7,
         }
         self._layer_config['format'] = {
             'type': InputType.List,