+++ lizmap/forms/table_manager.py
@@ -531,6 +531,19 @@ class TableManager:
 
                             layer_data[json_key] = trace[key]
 
+            if self.definitions.key() == 'formFilterLayers':
+                if version < LwcVersions.Lizmap_3_7:
+                    # We need to change keys to write in the legacy format
+                    if layer_data.get('end_field'):
+                        # Incompatible with this format, but we don't remove it just in case
+                        LOGGER.error(
+                            "A end_field is defined for the form filter. This is not compatible for this version of "
+                            "Lizmap Web Client"
+                        )
+                    if layer_data.get('start_field'):
+                        layer_data['field'] = layer_data.get('start_field')
+                        del layer_data['start_field']
+
             if export_legacy_single_row:
                 if self.definitions.key() == 'atlas':
                     layer_data['atlasEnabled'] = 'True'
@@ -638,6 +651,20 @@ class TableManager:
         return data
 
     @staticmethod
+    def _from_json_legacy_form_filter(data):
+        """ Read form filter and transform it if needed. """
+        for layer in data.get('layers'):
+            if layer.get('type') != 'numeric':
+                continue
+
+            if layer.get('field'):
+                # We upgrade from < 3.7 to 3.7 format
+                layer['start_field'] = layer['field']
+                del layer['field']
+
+        return data
+
+    @staticmethod
     def _from_json_legacy_dataviz(data):
         """Read legacy dataviz without the traces config."""
 
@@ -683,8 +710,11 @@ class TableManager:
 
         return data
 
-    def from_json(self, data):
-        """Load JSON into the table."""
+    def from_json(self, data: dict):
+        """Load JSON into the table.
+
+        :param data: The data for the given table to read from the CFG file : layers and general configuration
+        """
         if self.definitions.key() in (
             'locateByLayer',
             'loginFilteredLayers',
@@ -702,6 +732,9 @@ class TableManager:
 
         if self.definitions.key() == 'datavizLayers':
             data = self._from_json_legacy_dataviz(data)
+
+        if self.definitions.key() == 'formFilterLayers':
+            data = self._from_json_legacy_form_filter(data)
 
         config = data.get('config')
         # config: Union[dict, None]