+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased for the next version, so this won't be merged and no conflict
+
+* Add new "maximum field" for Lizmap Web Client ≥ 3.7 about the form filter with numeric values
+
 ## Unreleased
 
 * Remove notes about the Lizmap QGIS server plugin